[sonar] Explicit cast when loosing precision (teshsuite/).

+++ teshsuite/s4u/concurrent_rw/concurrent_rw.cpp
@@ -10,7 +10,7 @@ XBT_LOG_NEW_DEFAULT_CATEGORY(s4u_test, "Messages specific for this s4u test");
 static void host()
 {
   simgrid::s4u::Disk* disk = simgrid::s4u::this_actor::get_host()->get_disks().front(); // Disk1
-  int id                   = simgrid::s4u::this_actor::get_pid();
+  int id                   = static_cast<int>(simgrid::s4u::this_actor::get_pid());
   XBT_INFO("process %d is writing!", id);
   disk->write(4000000);
   XBT_INFO("process %d goes to sleep for %d seconds", id, id);
